+++ packages/app/src/pages/scene/deferred.tsx
@@ -10,7 +10,7 @@ import {
   GeometryData, PBRMaterial, ImageTexture,
   OrbitCamera, OrbitControls, Environment,
   Cursor,
-  DirectionalLight, PointLight, AmbientLight,
+  DirectionalLight, PointLight, AmbientLight, HemiLight,
   Data, PointLayer,
   makeBoxGeometry, makePlaneGeometry, makeSphereGeometry,
 } from '@use-gpu/workbench';
@@ -63,6 +63,11 @@ const lightData = [
     position: [2, 4.5, 2.5, 1],
     color: [0.3, 0.8, 1.0, 1],
   },
+  {
+    position: [5, 20, -3, 1],
+    direction: [-0.307, -1, 0.307, 1],
+    color: [0.25, 1.0, 0.1, 1],
+  },
 ];
 
 export const SceneDeferredPage: LC = () => {
@@ -93,9 +98,11 @@ export const SceneDeferredPage: LC = () => {
             <Loop converge={64}>
               <Pass lights shadows mode="deferred" ssao={2}>
                 <AmbientLight intensity={0.2} />
-                <DirectionalLight position={lightData[0].position} intensity={1}   color={lightData[0].color} shadowMap={SHADOW_MAP_DIRECTIONAL} />
-                <DirectionalLight position={lightData[1].position} intensity={0.5} color={lightData[1].color} shadowMap={SHADOW_MAP_DIRECTIONAL} />
-                <PointLight       position={lightData[2].position} intensity={100} color={lightData[2].color} shadowMap={SHADOW_MAP_POINT} />
+
+                <DirectionalLight {...lightData[0]} intensity={1}   shadowMap={SHADOW_MAP_DIRECTIONAL} />
+                <DirectionalLight {...lightData[1]} intensity={0.5} shadowMap={SHADOW_MAP_DIRECTIONAL} />
+                <PointLight       {...lightData[2]} intensity={100} shadowMap={SHADOW_MAP_POINT} />
+                <HemiLight        {...lightData[3]} intensity={100} shadowMap={SHADOW_MAP_POINT} />
 
                 <Environment preset="none">
                   <Scene>