+++ packages/scene/src/mesh.ts
@@ -3,18 +3,14 @@ import type { GPUGeometry } from '@use-gpu/core';
 
 import { memo, use, wrap } from '@use-gpu/live';
 
-import { FaceLayer } from '@use-gpu/workbench';
+import { FaceLayer, FaceLayerFlags } from '@use-gpu/workbench';
 
 import { Primitive } from './primitive';
 
 export type MeshProps = PropsWithChildren<{
   id?: number,
   mesh: GPUGeometry,
-  shaded?: boolean,
-  shadow?: boolean,
-  side?: 'front' | 'back' | 'both',
-  mode?: string,
-}>;
+}> & FaceLayerFlags;
 
 export const Mesh: LiveComponent<MeshProps> = memo((props: MeshProps) => {
   return (