+++ packages/workbench/src/light/types.ts
@@ -1,9 +1,10 @@
 import { mat4, vec4, vec3, vec2 } from 'gl-matrix';
 
 export const AMBIENT_LIGHT = 0;
-export const DIRECTIONAL_LIGHT = 1;
-export const DOME_LIGHT = 2;
+export const DOME_LIGHT = 1;
+export const DIRECTIONAL_LIGHT = 2;
 export const POINT_LIGHT = 3;
+export const HEMI_LIGHT = 4;
 
 export type Light = {
   into?: mat4 | null,